% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postupdate_blacklist.R
\name{postupdate_blacklist}
\alias{postupdate_blacklist}
\title{Update blacklist after the lecture}
\usage{
postupdate_blacklist(blacklist, newblacklist)
}
\arguments{
\item{blacklist}{The CDS blacklist: An object read by readxl package}

\item{newblacklist}{Names needed to be put on blacklist: An object read by readxl package}
}
\value{
Updated blacklist
}
\description{
To put someone broke the rules on the CDS blacklist
}
