% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_record.R
\name{search_record}
\alias{search_record}
\title{Search participation records}
\usage{
search_record(record, name)
}
\arguments{
\item{record}{An object read by readxl package}

\item{name}{A fixed format data frame with 2 columns}
}
\value{
Information of a specific applicant
}
\description{
If someone deliver a application for certificate, this function can be used to confirm their lectures participations.
}
