% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_record.R
\name{update_record}
\alias{update_record}
\title{Update the extant sign-in records}
\usage{
update_record(record, form, order)
}
\arguments{
\item{record}{An object read by readxl package}

\item{form}{A fixed format data frame with 2 columns}

\item{order}{The order for the new sign-in form}
}
\value{
A data frame used for storage
}
\description{
Update the record with a new sign-in form
}
\details{
The record and the sign-in form should be the fixed formats
}
