% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preupdate_blacklist.R
\name{preupdate_blacklist}
\alias{preupdate_blacklist}
\title{Update penalty blacklist before the lecture}
\usage{
preupdate_blacklist(blacklist, form)
}
\arguments{
\item{blacklist}{The CDS blacklist: An object read by readxl package}

\item{form}{Enrollment form of CDS lectures: An object read by readxl package}
}
\value{
Updated blacklist
}
\description{
To update the penalty blacklist when someone had been punished（removed from the enrollment form）
}
