% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_blacklist.R
\name{search_blacklist}
\alias{search_blacklist}
\title{Search blacklist and update enrollment form}
\usage{
search_blacklist(blacklist, form)
}
\arguments{
\item{blacklist}{The CDS blacklist: An object read by readxl package}

\item{form}{Enrollment form of CDS lectures: An object read by readxl package}
}
\value{
Updated enrollment form
}
\description{
To remove names on the blacklist from the enrollment form. There is no one need to be removed if no object returned.
}
